requirements.txt
numpy
pandas
PyTermGUI
requests
requests-toolbelt==1.0.0
rich
ruamel.yaml==0.17.32
scipy
tqdm